requirements.txt

Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.2.0
Flask-Login==0.6.3
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.1
